Modularize aliases system
